// modules/order/client-react/graphql/__generated__/addToCart.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { AddToCartInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: addToCart
// ====================================================

export interface addToCart {
  addToCart: boolean | null;
}

export interface addToCartVariables {
  input: AddToCartInput;
}