// modules/order/client-react/graphql/__generated__/editOrderDetail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { EditOrderDetailInput } from "./../../../../../packages/server/__generated__/globalTypes";

// ====================================================
// GraphQL mutation operation: editOrderDetail
// ====================================================

export interface editOrderDetail {
  editOrderDetail: boolean | null;
}

export interface editOrderDetailVariables {
  input?: EditOrderDetailInput | null;
}