// modules/order/client-react/graphql/__generated__/deleteOrderDetail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: deleteOrderDetail
// ====================================================

export interface deleteOrderDetail {
  deleteOrderDetail: boolean | null;
}

export interface deleteOrderDetailVariables {
  id: number;
}