// modules/order/client-react/graphql/__generated__/deleteOrder.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: deleteOrder
// ====================================================

export interface deleteOrder {
  deleteOrder: boolean | null;
}

export interface deleteOrderVariables {
  id: number;
}