// modules/order/client-react/graphql/__generated__/patchOrderState.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: patchOrderState
// ====================================================

export interface patchOrderState {
  patchOrderState: boolean | null;
}

export interface patchOrderStateVariables {
  orderId?: number | null;
  state?: string | null;
}