// modules/order/client-react/graphql/__generated__/patchAddress.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: patchAddress
// ====================================================

export interface patchAddress {
  patchAddress: boolean | null;
}

export interface patchAddressVariables {
  cartId?: number | null;
  addressId?: number | null;
}