// modules/order/client-react/graphql/__generated__/orderStatusMail.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: orderStatusMail
// ====================================================

export interface orderStatusMail {
  orderStatusMail: boolean | null;
}

export interface orderStatusMailVariables {
  orderId?: number | null;
  note?: string | null;
}